Rename sign in presentation directory
