+++ lib/features/auth/presentation/view/auth_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 import '../../../../ui/screens/sign_up/sign_up_screen.dart';
 import '../../../../ui/shared/extensions/build_context.dart';
-import '../../../sign_in/presentations/view/sign_in_screen.dart';
+import '../../../sign_in/presentation/view/sign_in_screen.dart';
 
 class AuthScreen extends StatelessWidget {
   const AuthScreen({super.key});
